Updated simple addition Job to be able to pass input range and number of partitions as parameters when calling REST endpoint.

+++ examples/src/main/java/examples/dev/jchejarla/springbatch/clustering/simplejob/MultiNodeExecutionTask.java
@@ -16,9 +16,9 @@ public class MultiNodeExecutionTask implements Tasklet {
         log.info("Executing MultiNode Task - START");
         StepExecution stepExecution =chunkContext.getStepContext().getStepExecution();
         ExecutionContext executionContext = stepExecution.getExecutionContext();
-        Integer start = executionContext.getInt("start");
-        Integer end = executionContext.getInt("end");
-        long result = (long)(end- start +1) * (start + end) /2;
+        Long start = executionContext.getLong("start");
+        Long end = executionContext.getLong("end");
+        long result = (end- start +1) * (start + end) /2;
         executionContext.put("result", result);
         log.info("Executing MultiNode Task - END");
         return RepeatStatus.FINISHED;
